To-do list
a very simple todo list that i whipped up in the lab. First time using js, html and css together. Followed a little bit of a tutorial when scripting because i didnt know exactly how html and javascript interact (i.e methods and functions), however, most html and css i did myself (excluding the parts i added from the tut to work with the js). Took the better half of a day, was watching a css tut by Bro Code as a refresher, but I learned a lot from this project, including local data storage and recalling that data. Left a couple comments in there too. Hope to commit a lot more projects in the future.